Identity Server workflow to run on any change
The workflow was copied from 7.1 to 7.0 in
https://github.com/DuendeSoftware/products/pull/1979, but the structure
of the repository does not match the structure of the repository from
which it was copied.

For example, the path trigger indicates that /identity-server as a
directory exists, it does not. The Identity Server source code exists in
src/IdentityServer and the tests exist in /test.

It's easier to have the workflow run on any change.